1595: pared down Factotum landing page (#1636)
* 1595: pared down Factotum landing page

* Move stat logic and cards into separate sources
