+++ dashboard/urls.py
@@ -239,6 +239,8 @@ urlpatterns = [
         name="link_habitsandpractices",
     ),
     path("get_data/", views.get_data, name="get_data"),
+    path("statistics/", views.Statistics.as_view(), name="statistics"),
+    path("visualizations/", views.Visualizations.as_view(), name="visualizations"),
     path("bulk_documents/", views.BulkDocuments.as_view(), name="bulk_documents"),
     path(
         "bulk_rawcategory/",
