+++ templates/core/base.html
@@ -73,6 +73,26 @@
                     Get Data
                 </a>
             </li>
+            <li class="nav-item
+                            {% if '/statistics/' in request.path %}
+                                active
+                            {% endif %}">
+                <a class="nav-link"
+                   href="{% url 'statistics' %}"
+                   title="Statistics">
+                    Statistics
+                </a>
+            </li>
+            <li class="nav-item
+                            {% if '/visualizations/' in request.path %}
+                                active
+                            {% endif %}">
+                <a class="nav-link"
+                   href="{% url 'visualizations' %}"
+                   title="Visualizations">
+                    Visualizations
+                </a>
+            </li>
             {% if user.is_authenticated %}
                 <li class="nav-item
                             {% if '/datasources/' in request.path %}
