+++ templates/dashboard/index.html
@@ -1,145 +1,9 @@
 {% extends 'core/base.html' %}
-{% load staticfiles %}
-{% load humanize %}
-{% load tags %}
 
 {% block content %}
     <h1 class="col-sm-12 p-2"><span class="fa fa-wrench" title="factotum" aria-hidden="true"></span>
         Welcome to Factotum
     </h1>
 
-    <div class="card-deck">
-        <div class="card text-center bg-light">
-            <div class="card-body">
-                <h3 class="card-title text-muted">Documents</h3>
-                <div class="card-text font-weight-bold display-4">
-                    {% if datadocument_count > 999999 %}
-                        {{ datadocument_count | intword }}
-                    {% else %}
-                        {{ datadocument_count | intcomma }}
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-        <div class="card text-center bg-light">
-            <div class="card-body">
-                <h3 class="card-title text-muted">Products</h3>
-                <div class="card-text font-weight-bold display-4">
-                    {% if product_count > 999999 %}
-                        {{ product_count | intword }}
-                    {% else %}
-                        {{ product_count | intcomma }}
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-        <div class="card text-center bg-light">
-            <div class="card-body">
-                <h3 class="card-title text-muted">Extracted Chemicals</h3>
-                <div class="card-text font-weight-bold display-4">
-                    {% if chemical_count > 999999 %}
-                        {{ chemical_count | intword }}
-                    {% else %}
-                        {{ chemical_count | intcomma }}
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div class="card-deck pt-2 mt-2">
-        <div class="card text-center bg-light">
-            <div class="card-body">
-                <h3 class="card-title text-muted">Products Linked To PUCs</h3>
-                <div name="product_with_puc_count" class="card-text font-weight-bold display-4">
-                    {% if product_with_puc_count > 999999 %}
-                        {{ product_with_puc_count | intword }}
-                    {% else %}
-                        {{ product_with_puc_count | intcomma }}
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-        <div class="card text-center bg-light">
-            <div class="card-body">
-                <h3 class="card-title text-muted">Curated Chemical Records</h3>
-                <div name="chemical_records" class="card-text font-weight-bold display-4">
-                    {% if curated_chemical_count > 999999 %}
-                        {{ curated_chemical_count | intword }}
-                    {% else %}
-                        {{ curated_chemical_count | intcomma }}
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-        <div class="card text-center bg-light">
-            <div class="card-body">
-                <h3 class="card-title text-muted">Unique DTXSIDs</h3>
-                <div class="card-text font-weight-bold display-4" name="dsstox">
-                    {% if dsstox_sid_count > 999999 %}
-                        {{ dsstox_sid_count | intword }}
-                    {% else %}
-                        {{ dsstox_sid_count | intcomma }}
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div class="mt-5">
-        <table class="table table-striped dataTable no-footer table-sm"
-               style="width:100%"
-               id="grouptype_table"
-        >
-            <tr>
-                <td class="text-center" colspan="4"><i class="fas fa-spinner fa-pulse"></i></td>
-            </tr>
-        </table>
-    </div>
-
-    <div class="row mt-5 mb-5">
-        <svg id="collapsibletree" width="100%"></svg>
-    </div>
-
-    <h4 class="mt-5">Products Linked to <a href="/pucs">PUCs</a></h4>
-
-    {# Formulation PUCs    #}
-    <div class="row">
-        <div class="col-5 p-0">
-            {% show_bubble_puc_legend formulation_pucs "FO" False %}
-        </div>
-        <div class="col-7">
-            <svg id="nestedcircles_FO" class="nestedcircles" width="100%" height="500"></svg>
-        </div>
-    </div>
-
-    {# Article PUCs    #}
-    <div class="row">
-        <div class="col-5 p-0">
-            {% show_bubble_puc_legend article_pucs "AR" False %}
-        </div>
-        <div class="col-7">
-            <svg id="nestedcircles_AR" class="nestedcircles" width="100%" height="500"></svg>
-        </div>
-    </div>
-
-    <div class="row mt-5 mb-5">
-        <h4>SIDs per Group Type</h4>
-
-        <div id="venn"></div>
-    </div>
-{% endblock %}
-
-{% block css %}
-    <link href="{% static 'css/dashboard/index.css' %}" rel="stylesheet">
-{% endblock %}
-
-{% block js %}
-    <script src="{% static 'js/d3.v5.min.js' %}"></script>
-    <script src="{% static 'js/d3-selection.v1.min.js' %}"></script>
-    <script src="{% static 'js/bubble_chart.js' %}"></script>
-    <script src="{% static 'js/bubble_puc_legend.js' %}"></script>
-    <script src="{% static 'js/collapsible_tree.js' %}"></script>
-    <script src="{% static 'js/dashboard/venn.min.js' %}"></script>
-    <script src="{% static 'js/dashboard/index.js' %}"></script>
+    {% include "statistics/stat_cards.html" %}
 {% endblock %}
